make fifo queue default (again) as fringe for debugging.
